+++ src/earth.c
@@ -15,6 +15,11 @@
 /// \cond PRIVATE
 #define __NOVAS_INTERNAL_API__    ///< Use definitions meant for internal use by SuperNOVAS only
 #include "novas.h"
+
+#define ERAD                NOVAS_GRS80_RADIUS
+#define EF                  NOVAS_GRS80_FLATTENING
+#define ANGVEL              NOVAS_EARTH_ANGVEL
+#define ERAD_AU             ( ERAD / NOVAS_AU )
 /// \endcond
 
 #if __cplusplus